requirements.txt
django==2.2.2
gunicorn==20.1.0
pre-commit==2.15.0
requests==2.26.0
djangorestframework==3.12.4
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
